implementnación del primer punto de la tercera parte
se hace movimiento del menú de registrar votos hacia la clase main, se crea un atributo votos en la clase candidato, se crea el método contarVotos en la clase métodos en cual consiste en el punto 1. Qué candidato ganó, su nombre y cédula
